pyproject.toml
[project]
name = "educational-video-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imageio-ffmpeg>=0.6.0",
    "manim>=0.19.0",
    "moviepy>=2.2.1",
    "numpy>=2.3.4",
    "openai>=2.7.1",
    "pydub>=0.25.1",
    "pydantic>=2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "streamlit>=1.51.0",
    "typer>=0.20.0",
]

[project.scripts]
video = "educational_video_generation.cli:main"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E203"]

[tool.ruff.lint.isort]
known-first-party = ["educational_video_generation"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
skip-magic-trailing-comma = false

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]
